refactor(server): rename lambda.ts and node.ts to *-converters.ts
