+++ packages/server-apigateway/src/index.ts
@@ -1,16 +1,6 @@
 /*
- *  Copyright 2021 Amazon.com, Inc. or its affiliates. All Rights Reserved.
- *
- *  Licensed under the Apache License, Version 2.0 (the "License").
- *  You may not use this file except in compliance with the License.
- *  A copy of the License is located at
- *
- *   http://aws.amazon.com/apache2.0
- *
- *  or in the "license" file accompanying this file. This file is distributed
- *  on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either
- *  express or implied. See the License for the specific language governing
- *  permissions and limitations under the License.
+ * Copyright Amazon.com, Inc. or its affiliates. All Rights Reserved.
+ * SPDX-License-Identifier: Apache-2.0
  */
 
-export { convertEvent, convertVersion1Response, convertVersion2Response } from "./lambda";
+export { convertEvent, convertVersion1Response, convertVersion2Response } from "./lambda-converters";
