+++ packages/server-node/src/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-export { convertRequest, writeResponse } from "./node";
+export { convertRequest, writeResponse } from "./node-http-converters";
